+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v 16 "..." (WIP)
+- new: tracking-URL-popup
 - fix: context menu in tab overview
 - fix: onRenderProcessGone
 - updated: translations